refactor: narrow types.ts to pure type declarations
